TA-3769: Move Pacman related commands out of Studio module
